demo security update

+++ app/presenters/AdminPresenter.php
@@ -2,6 +2,7 @@
 
 namespace App;
 
+use App;
 use Cntrl;
 use Entity;
 use Kdyby;
@@ -116,6 +117,11 @@ class AdminPresenter extends BasePresenter {
 	 * @param $id
 	 */
 	public function colorSucceeded($button, $id) {
+		if (!$this->editable()) {
+			$this->presenter->flashMessage('Myslím to vážně, editovat opravdu **ne**můžete!', 'danger');
+			$this->redirect('this');
+			return;
+		}
 		$vals = $button->getForm()->getValues();
 		$newColor = preg_replace('<#>', '', $vals['color' . $id]);
 		if (ctype_xdigit($newColor) && (strlen($newColor) == 6 || strlen($newColor) == 3)) {
@@ -151,7 +157,11 @@ class AdminPresenter extends BasePresenter {
 	}
 
 	public function handleDelete($id) {
-		//TODO: pro demo uživatele zamezit možnost mazání
+		if (!$this->editable()) {
+			$this->presenter->flashMessage('Myslím to vážně, editovat (ani mazat) opravdu **ne**můžete!', 'danger');
+			$this->redirect('this');
+			return;
+		}
 		try {
 			$this->posts->delete($this->posts->findOneBy(array('id' => $id)));
 			$this->flashMessage('Článek byl úspěšně smazán.', 'success');
@@ -162,6 +172,11 @@ class AdminPresenter extends BasePresenter {
 	}
 
 	public function handleDeleteTag($tag_id) {
+		if (!$this->editable()) {
+			$this->presenter->flashMessage('Myslím to vážně, editovat (ani mazat) opravdu **ne**můžete!', 'danger');
+			$this->redirect('this');
+			return;
+		}
 		try {
 			$this->tags->delete($this->tags->findOneBy(array('id' => $tag_id)));
 			$this->flashMessage('Tag byl úspěšně smazán.', 'success');
@@ -172,6 +187,11 @@ class AdminPresenter extends BasePresenter {
 	}
 
 	public function handleDeleteUser($user_id) {
+		if (!$this->editable()) {
+			$this->presenter->flashMessage('Myslím to vážně, editovat (ani mazat) opravdu **ne**můžete!', 'danger');
+			$this->redirect('this');
+			return;
+		}
 		try {
 			$this->users->delete($this->users->findOneBy(['id' => $user_id]));
 			$this->flashMessage('Uživatel byl úspěšně smazán.', 'success');
@@ -182,6 +202,11 @@ class AdminPresenter extends BasePresenter {
 	}
 
 	public function handleRegenerate($tag_id) {
+		if (!$this->editable()) {
+			$this->presenter->flashMessage('Myslím to vážně, editovat opravdu **ne**můžete!', 'danger');
+			$this->redirect('this');
+			return;
+		}
 		try {
 			$tag = $this->tags->findOneBy(array('id' => $tag_id));
 			$tag->color = substr(md5(rand()), 0, 6); //Short and sweet
@@ -222,12 +247,21 @@ class AdminPresenter extends BasePresenter {
 	}
 
 	public function handleDeletePicture($id) {
+		if (!$this->editable()) {
+			$this->presenter->flashMessage('Myslím to vážně, editovat (ani mazat) opravdu **ne**můžete!', 'danger');
+			$this->redirect('this');
+			return;
+		}
 		$picture = $this->pictures->findOneBy(['id' => $id]);
 		@unlink(__DIR__ . '/../../www/uploads/' . $picture->uuid . DIRECTORY_SEPARATOR . $picture->name);
 		@rmdir(__DIR__ . '/../../www/uploads/' . $picture->uuid);
 		$this->pictures->delete($picture);
 		$this->flashMessage('Obrázek byl úspěšně smazán.', 'success');
 		$this->redirect('this');
+	}
+
+	private function editable() {
+		return $this->user->isAllowed('Admin', App\Authorizator::EDIT) ? TRUE : FALSE;
 	}
 
 }
